Make all img tags responsive to container size.
I had updated individual posts to have `width="100%"`, which of course
did nothing for the posts I missed. I attempted to fix this in sweeping
fashion by updating the root styelsheet to set a max-width for all
images. This shouldn't affect small images at all, but will make sure
very large images are resized when appropriate for the container.